Add test to check if atom is installed

+++ setup/atom.sh
@@ -1,10 +1,14 @@
-apm stars --install
+if which apm >/dev/null; then
+	apm stars --install
 
-rm ~/.atom/keymap.cson
-ln -s ~/dotfiles/atom/keymap.cson ~/.atom/keymap.cson
+	rm ~/.atom/keymap.cson
+	ln -s ~/dotfiles/atom/keymap.cson ~/.atom/keymap.cson
 
-rm ~/.atom/config.cson
-ln -s ~/dotfiles/atom/config.cson ~/.atom/config.cson
+	rm ~/.atom/config.cson
+	ln -s ~/dotfiles/atom/config.cson ~/.atom/config.cson
 
-rm ~/.atom/styles.less
-ln -s ~/dotfiles/atom/styles.less ~/.atom/styles.less
+	rm ~/.atom/styles.less
+	ln -s ~/dotfiles/atom/styles.less ~/.atom/styles.less
+else
+	echo "Atom not installed"
+fi
